Re-namespace ClearInfo/MaskingInfo/DepthInfo
