+++ osu.Framework/Graphics/Rendering/ClearInfo.cs
@@ -1,11 +1,9 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-#nullable disable
-
 using osuTK.Graphics;
 
-namespace osu.Framework.Graphics.OpenGL
+namespace osu.Framework.Graphics.Rendering
 {
     /// <summary>
     /// Information for how the current frame buffer should be cleared.
