+++ osu.Framework/Graphics/Rendering/DepthInfo.cs
@@ -1,12 +1,10 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-#nullable disable
-
 using System;
 using osuTK.Graphics.ES30;
 
-namespace osu.Framework.Graphics.OpenGL
+namespace osu.Framework.Graphics.Rendering
 {
     /// <summary>
     /// Information for how depth should be handled.
