+++ osu.Framework/Graphics/Rendering/DepthValue.cs
@@ -1,11 +1,9 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-#nullable disable
-
 using System.Runtime.CompilerServices;
 
-namespace osu.Framework.Graphics.OpenGL
+namespace osu.Framework.Graphics.Rendering
 {
     /// <summary>
     /// The depth value used to draw 2D objects to the screen.
